+++ src/components/TodoList/TodoListItem/TodoListItem.tsx
@@ -1,6 +1,6 @@
-import { TodoListItemTypes } from '../../context/TodoContext';
-import styles from './TodoList.module.css';
-import TodoContext from '../../context/TodoContext';
+import { TodoListItemTypes } from '../../../context/TodoContext';
+import styles from './TodoListItem.module.css';
+import TodoContext from '../../../context/TodoContext';
 import { useContext } from 'react';
 
 interface TodoListItemProps {
